+++ src/graphql.tsx
@@ -124,6 +124,11 @@ export default function graphql<
       private querySubscription: Subscription;
       private previousData: any = {};
       private lastSubscriptionData: any;
+      private refetcherQueue: {
+        args: any;
+        resolve: (value?: {} | PromiseLike<{}>) => void;
+        reject: (reason?: any) => void;
+      };
 
       // calculated switches to control rerenders
       private shouldRerender: boolean;
@@ -155,6 +160,11 @@ export default function graphql<
 
         if (!this.shouldSkip(this.props)) {
           this.subscribeToQuery();
+          // call any stacked refetch functions
+          if (this.refetcherQueue) {
+            const { args, resolve, reject } = this.refetcherQueue;
+            this.queryObservable.refetch(args).then(resolve).catch(reject);
+          }
         }
       }
 
@@ -215,8 +225,11 @@ export default function graphql<
         if (this.type === DocumentType.Query) {
           // Recycle the query observable if there ever was one.
           if (this.queryObservable) {
-            this.getQueryRecycler().recycle(this.queryObservable);
-            delete this.queryObservable;
+            const recycler = this.getQueryRecycler();
+            if (recycler) {
+              recycler.recycle(this.queryObservable);
+              delete this.queryObservable;
+            }
           }
 
           // It is critical that this happens prior to recyling the query
@@ -234,12 +247,15 @@ export default function graphql<
       }
 
       getQueryRecycler() {
-        return this.context.getQueryRecycler(GraphQL);
+        return (
+          this.context.getQueryRecycler &&
+          this.context.getQueryRecycler(GraphQL)
+        );
       }
 
-      getClient(): ApolloClient {
+      getClient(props): ApolloClient {
         if (this.client) return this.client;
-        const { client } = mapPropsToOptions(this.props);
+        const { client } = mapPropsToOptions(props);
 
         if (client) {
           this.client = client;
@@ -323,9 +339,9 @@ export default function graphql<
         this.createQuery(opts);
       }
 
-      createQuery(opts: QueryOpts) {
+      createQuery(opts: QueryOpts, props: any = this.props) {
         if (this.type === DocumentType.Subscription) {
-          this.queryObservable = this.getClient().subscribe(
+          this.queryObservable = this.getClient(props).subscribe(
             assign(
               {
                 query: document,
@@ -337,10 +353,12 @@ export default function graphql<
           // Try to reuse an `ObservableQuery` instance from our recycler. If
           // we get null then there is no instance to reuse and we should
           // create a new `ObservableQuery`. Otherwise we will use our old one.
-          const queryObservable = this.getQueryRecycler().reuse(opts);
+          const recycler = this.getQueryRecycler();
+          let queryObservable = null;
+          if (recycler) queryObservable = recycler.reuse(opts);
 
           if (queryObservable === null) {
-            this.queryObservable = this.getClient().watchQuery(
+            this.queryObservable = this.getClient(props).watchQuery(
               assign(
                 {
                   query: document,
@@ -364,7 +382,7 @@ export default function graphql<
 
         // if we skipped initially, we may not have yet created the observable
         if (!this.queryObservable) {
-          this.createQuery(opts);
+          this.createQuery(opts, props);
         }
 
         if (this.queryObservable._setOptionsNoResult) {
@@ -403,7 +421,7 @@ export default function graphql<
           opts.fetchPolicy = 'cache-first'; // ignore force fetch in SSR;
         }
 
-        const observable = this.getClient().watchQuery(
+        const observable = this.getClient(this.props).watchQuery(
           assign({ query: document }, opts),
         );
         const result = observable.currentResult();
@@ -498,7 +516,7 @@ export default function graphql<
         if (typeof opts.variables === 'undefined') delete opts.variables;
 
         (opts as any).mutation = document;
-        return this.getClient().mutate(opts as any) as Promise<
+        return this.getClient(this.props).mutate(opts as any) as Promise<
           ApolloQueryResult<TResult>
         >;
       }
@@ -561,6 +579,15 @@ export default function graphql<
           } else {
             assign(data, currentResult.data);
             this.previousData = currentResult.data;
+          }
+
+          // handle race condition where refetch is called on child mount
+          if (!this.querySubscription) {
+            (data as QueryProps).refetch = args => {
+              return new Promise((r, f) => {
+                this.refetcherQueue = { resolve: r, reject: f, args };
+              });
+            };
           }
         }
         return data as QueryProps & TResult;