+++ Changelog.md
@@ -1,9 +1,20 @@
 # Change log
 
 ### vNext
-- Feature: Support apollo-client 2.0
-- Fix: Scope query recyclers by client [#876](https://github.com/apollographql/react-apollo/pull/876)
+- Added notifyOnNetworkStatusChange to QueryOpts and MutationOpts Typesccript definitions [#1034](https://github.com/apollographql/react-apollo/pull/1034)
+
+### 1.4.15
+- Fix: handle calling refetch in child componentDidMount
+- Fix: ensure options gets up to date props [#1025](https://github.com/apollographql/react-apollo/pull/1005)
+- Fix: ensure queryRecycler exists before using it
 - MockNetworkInterface match mock requests regardless of variable order [#973](https://github.com/apollographql/react-apollo/pull/973)
+- Allow to pass removeTypenames to MockedProvider [#1001](https://github.com/apollographql/react-apollo/pull/1001)
+
+### 1.4.14
+- Fix: Scope query recyclers by client [#876](https://github.com/apollographql/react-apollo/pull/876)
+
+### 1.4.13 [DEPRECATED]
+- Support apollo-client 2.0
 
 ### 1.4.12
 - Fix: fix issue with bad deploy